Modified randomSeed, now uses unsigned long.

+++ hardware/arduino/avr/cores/arduino/Arduino.h
@@ -239,7 +239,7 @@ void noTone(uint8_t _pin);
 // WMath prototypes
 long random(long);
 long random(long, long);
-void randomSeed(unsigned int);
+void randomSeed(unsigned long);
 long map(long, long, long, long, long);
 
 #endif
